Converted wjelement
First iteration took a lot of prompting to get it on track translating cmake files. It eventually got there and adding most of what was needed to the rules.md.

Second iteration went way smoother. a testiment to the improved rules. something that was a bit weird was that the new rules.md it output was substantially worse than the first. it added a bunch of new practices that it didnt use and werent great practices. Not totally sure how to address that currently. will see if this problem persists in other builds.

Really need to improve the prompt to just tell it "if you need these dependencies, use these versions". It keeps using the super outdated versions of rules_cc and platforms.