First iteration of converting parson v2
Mostly went fine until I realized that 2 tests werent passing. This is due to the tests trying to write out a file in the sandbox which doesnt have write permissions. Not totally sure how to handle this case. First conversion also had this issue. it was missed because bazel doesnt show test output of passing tests and the test returns a 0 error code even when it fails. Updated rules to always include `--test_output=all` to handle this.