First iteration after converting cmocka
This was a challenging conversion. cmocka does some weird things in the tests. it performs source file inclusion and reports non-zero return code for failing tests that are expected to fail. had to come up with a workaround for both and took a lot of instructing the agent to get it to work.

Even after it got the solution, it still spun wheels for a bit trying to update the tests and when running the tests.

I think the process improvements section needs better formatting overall. Not sure what I want to improve about it yet. It seems unorganized and not really giving clear actions to take in various circumstances

Lastly, the two new artifacts definitely need better structure. The conversion log I want to be way more detailed and the module index should pretty much just be a summary of the public interface of the module.