cmocka conversion 2
This was a test of using the conversion log to aid in conversion and even though i wasnt super happy with the original conversion log, it worked. That conversion didnt need any manual intervention. The only time I intervened was to prompt it to not rerun the tests after it just ran them, but that wasnt necessary to get it to complete the conversion. I will note that it didnt put the BUILD file in the tests folder for tests but did scope the private library for the test folder. but aside from that fairly minor nit, that was a complete success