Converted cmp
This conversion was pretty straightforward. It was the first time that I tried converting something that had a dependency on another project. The module-index.md idea worked. There were two super impressive things. cmocka requires including things like stdint.h before including its header to give you the ability to redefine those, and the cmp was using a very old version of cmocka and didnt have this. the agent figured it out on its own and fixed it. it also updated the api to new cmocka api and made all the tests run.